cmd/snap-update-ns: move system profile code to system.go
This will allow easier comparison with upcoming user.go
